+++ src/jawl_voicecompanion/__init__.py
@@ -2,15 +2,18 @@
 
 from .gateway import TextGateway
 from .hostos_policy import HostOSPolicy
+from .hostos_tools import HostOSExecutor, ToolSpec
 from .models import AccessLevel, RiskClass, ToolRequest
 from .arbiter import TurnArbiter, TurnPriority
 
 __all__ = [
     "AccessLevel",
     "HostOSPolicy",
+    "HostOSExecutor",
     "RiskClass",
     "TextGateway",
     "ToolRequest",
+    "ToolSpec",
     "TurnArbiter",
     "TurnPriority",
 ]